refactor(columnar): align NumericColumnValues.valueCount() to long
Aligns the write-path cursor with the read side, where valueCount(rank)
and firstOrdinal(int) already return long. Per-document counts cannot
overflow int today, but consistent types reduce the surface that would
need updating if Lucene ever widens docValueCount() upstream.

+++ libs/columnar/src/main/java/org/elasticsearch/columnar/ColumNARDocValuesConsumer.java
@@ -172,7 +172,7 @@ final class ColumNARDocValuesConsumer extends DocValuesConsumer {
             }
 
             @Override
-            public int valueCount() {
+            public long valueCount() {
                 return current.values.valueCount();
             }
